Feat: new connect button + deprecate mobile pairing (#2544)
* Feat: new connect button + deprecate mobile pairing

* 15 November

* Text edits
