+++ src/components/common/ConnectWallet/index.tsx
@@ -1,12 +1,12 @@
 import type { ReactElement } from 'react'
 import useWallet from '@/hooks/wallets/useWallet'
 import AccountCenter from '@/components/common/ConnectWallet/AccountCenter'
-import ConnectionCenter from '@/components/common/ConnectWallet/ConnectionCenter'
+import ConnectWalletButton from './ConnectWalletButton'
 
 const ConnectWallet = (): ReactElement => {
   const wallet = useWallet()
 
-  return wallet ? <AccountCenter wallet={wallet} /> : <ConnectionCenter />
+  return wallet ? <AccountCenter wallet={wallet} /> : <ConnectWalletButton />
 }
 
 export default ConnectWallet
