+++ src/components/common/ConnectWallet/styles.module.css
@@ -1,8 +1,3 @@
-.connectedContainer {
-  display: flex;
-  align-items: center;
-}
-
 .buttonContainer {
   display: flex;
   align-items: center;
@@ -45,22 +40,4 @@
 
 .row:last-of-type {
   border-bottom: 1px solid var(--color-border-light);
-}
-
-.pairingDetails {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: var(--space-2);
-}
-
-@media (max-width: 599.95px) {
-  .buttonContainer {
-    transform: scale(0.8);
-  }
-
-  .notConnected,
-  .pairingDetails {
-    display: none;
-  }
 }
