+++ src/components/new-safe/create/steps/ConnectWalletStep/index.tsx
@@ -1,8 +1,6 @@
 import { useEffect, useState } from 'react'
-import { Box, Button, Grid, Typography } from '@mui/material'
+import { Box, Button } from '@mui/material'
 import useWallet from '@/hooks/wallets/useWallet'
-import { useCurrentChain } from '@/hooks/useChains'
-import { isPairingSupported } from '@/services/pairing/utils'
 
 import type { NewSafeFormData } from '@/components/new-safe/create'
 import type { StepRenderProps } from '@/components/new-safe/CardStepper/useCardStepper'
@@ -10,15 +8,11 @@ import useSyncSafeCreationStep from '@/components/new-safe/create/useSyncSafeCre
 import layoutCss from '@/components/new-safe/create/styles.module.css'
 import useConnectWallet from '@/components/common/ConnectWallet/useConnectWallet'
 import KeyholeIcon from '@/components/common/icons/KeyholeIcon'
-import PairingDescription from '@/components/common/PairingDetails/PairingDescription'
-import PairingQRCode from '@/components/common/PairingDetails/PairingQRCode'
 import { usePendingSafe } from '../StatusStep/usePendingSafe'
 
 const ConnectWalletStep = ({ onSubmit, setStep }: StepRenderProps<NewSafeFormData>) => {
   const [pendingSafe] = usePendingSafe()
   const wallet = useWallet()
-  const chain = useCurrentChain()
-  const isSupported = isPairingSupported(chain?.disabledWallets)
   const handleConnect = useConnectWallet()
   const [, setSubmitted] = useState(false)
   useSyncSafeCreationStep(setStep)
@@ -34,31 +28,12 @@ const ConnectWalletStep = ({ onSubmit, setStep }: StepRenderProps<NewSafeFormDat
   }, [onSubmit, wallet, pendingSafe])
 
   return (
-    <>
-      <Box className={layoutCss.row}>
-        <Grid container spacing={3}>
-          <Grid item xs={12} md={6} display="flex" flexDirection="column" alignItems="center" gap={2}>
-            <Box width={100} height={100} display="flex" alignItems="center" justifyContent="center">
-              <KeyholeIcon />
-            </Box>
-
-            <Button onClick={handleConnect} variant="contained" size="stretched" disableElevation>
-              Connect
-            </Button>
-          </Grid>
-
-          {isSupported && (
-            <Grid item xs={12} md={6} display="flex" flexDirection="column" alignItems="center" gap={2}>
-              <PairingQRCode />
-              <Typography variant="h6" fontWeight="700">
-                Connect to {'Safe{Wallet}'} mobile
-              </Typography>
-              <PairingDescription />
-            </Grid>
-          )}
-        </Grid>
-      </Box>
-    </>
+    <Box className={layoutCss.row} display="flex" alignItems="center" gap={3}>
+      <KeyholeIcon />
+      <Button onClick={handleConnect} variant="contained" size="stretched" disableElevation>
+        Connect wallet
+      </Button>
+    </Box>
   )
 }
