+++ src/services/pairing/QRModal.tsx
@@ -8,6 +8,7 @@ import { StoreHydrator } from '@/store'
 import { AppProviders } from '@/pages/_app'
 import { PAIRING_MODULE_LABEL } from '@/services/pairing/module'
 import css from './styles.module.css'
+import PairingDeprecationWarning from '@/components/common/PairingDetails/PairingDeprecationWarning'
 
 const WRAPPER_ID = 'safe-mobile-qr-modal'
 const QR_CODE_SIZE = 200
@@ -75,6 +76,7 @@ const Modal = ({ uri, cb }: { uri: string; cb: () => void }) => {
             </IconButton>
           </DialogTitle>
           <DialogContent sx={{ display: 'flex', flexDirection: 'column', alignItems: 'center', gap: 1 }}>
+            <PairingDeprecationWarning />
             <PairingQRCode size={QR_CODE_SIZE} />
             <br />
             <PairingDescription />
